Removes `dynamics/dynamics.py`
Instead, adds `dynamics/dynamics_base.py` which contains an ABC for the
l2hmc dynamics object.

An example subclass can be found in `dynamics/gauge_dynamics.py`
